requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql[fastapi]
psycopg[binary,pool]
pylint
ipython
